MinGW-w64 support
For now, SGC must be used, but `configure' does not select it
automatically.

Also, support Cygwin (in addition to MSYS) as a build environment
when using MinGW compilers. Since I build in a Cygwin environment
(which seemed to be the easiest way to get MinGW-w64 gcc), I use

  ../configure --host=x86_64-w64-mingw32 --enable-sgc

+++ src/racket/include/scheme.h
@@ -244,8 +244,8 @@ typedef _int64 mzlonglong;
 typedef unsigned _int64 umzlonglong;
 #else
 # if defined(NO_LONG_LONG_TYPE) || defined(SIXTY_FOUR_BIT_INTEGERS)
-typedef long mzlonglong;
-typedef unsigned long umzlonglong;
+typedef intptr_t mzlonglong;
+typedef uintptr_t umzlonglong;
 # else
 typedef long long mzlonglong;
 typedef unsigned long long umzlonglong;
